compress: add security warning to docs
The compression standard may be insecure without taking special care to
flush the dictionary on stanza boundaries.  This must be done on both
the client and the server, and means that compression can't be used
safely unless all implementations behave correctly.  Because of this,
add a warning in the documentation to try and discourage use of this
package.